LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-never-https-http.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <script src="../resources/referrer-test.js"></script>
</head>
<body>
    <script>
        description("Test that 'no-referrer' works correctly from HTTPS to HTTP.");
        expectNoReferrer(ReferrerTest.NO_REFERRER, ReferrerTest.HTTPS, ReferrerTest.HTTP);
    </script>
</body>
</html>
